// src/types/lyrics.tsx
export type LyricsList = {
    name: string;
    title: string;
    album: string;
    language: string;
    year: string;
    description: string;
    cover: string;
}

export type Lyrics = {
    title: string;
    album: string;
    language: string;
    year: string;
    description: string;
    cover: string;
    content: string;
}